Add legacy function by accident.
libc: Fix absolute_timespec_from_timespec overflow.

A legacy library has uncovered an overflow in this function.
absolute_timespec_from_timespec blindly adds a relative timespec to
the current one, ignoring the possibility of tv_sec overflowing.

This was produced by calling pthread_cond_timedwait_relative_np with a
filled (to maximum values) timespec to create a faux 'infinite timeout'.

This is fixed by clamping the value as high as we can determine is safe.

Change-Id: I36ee2028dcf93e0d03246a6f0dd115cd3d11eaa3